# By minggo (12) and others
# Via minggo (18) and others
* develop: (78 commits)
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Update CHANGELOG[ci skip]
  Update AUTHORS [ci skip]
  fix bug: cccolor to jsval function has error in jsb
  [ci skip]
  [ci skip]
  [ci skip]
  [ci skip]
  [ci skip]
  C++11: NULL -> nullptr
  [ci skip]
  Fix: UserDefault::createXMLFile with correct XML definition.
  [ci skip]
  [ci skip]
  [ci skip]
  Fix bug: rename const with under line seperator
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Fix bug: Z fighting on TransitionScenePageTurn
  Fixed bugs
  Modify reader and fixed bugs
  ...

Conflicts:
	cocos/editor-support/cocostudio/CCSGUIReader.cpp

+++ cocos/2d/ZipUtils.cpp
@@ -323,7 +323,7 @@ bool ZipUtils::isCCZFile(const char *path)
 
 bool ZipUtils::isCCZBuffer(const unsigned char *buffer, long len)
 {
-    if (len < sizeof(struct CCZHeader))
+    if (static_cast<size_t>(len) < sizeof(struct CCZHeader))
     {
         return false;
     }
